Update to a daily build of .NET 6 RC1
Update to a daily build of RC1 that fixes issues with WebApplicationBuilder and MVC controllers with WebApplicationFactory<T>.
