+++ Directory.Packages.props
@@ -3,8 +3,8 @@
     <PackageVersion Include="coverlet.msbuild" Version="3.1.0" />
     <PackageVersion Include="MartinCostello.Logging.XUnit" Version="0.1.2" />
     <PackageVersion Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.18.0" />
-    <PackageVersion Include="Microsoft.AspNetCore.AzureAppServices.HostingStartup" Version="6.0.0-preview.6.21355.2" />
-    <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.0-preview.6.21355.2" />
+    <PackageVersion Include="Microsoft.AspNetCore.AzureAppServices.HostingStartup" Version="6.0.0-rc.1.21406.2" />
+    <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.0-rc.1.21406.2" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="16.10.0" />
     <PackageVersion Include="Microsoft.Playwright" Version="1.13.0" />
     <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.0.0" />
