+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "6.0.100-preview.6.21355.2",
+    "version": "6.0.100-rc.1.21406.9",
     "allowPrerelease": false,
     "rollForward": "latestMajor"
   }
