+++ tests/Website.Tests/Integration/TestServerFixture.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Martin Costello, 2016. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-using System.IO;
 using MartinCostello.Logging.XUnit;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
@@ -15,7 +14,7 @@ namespace MartinCostello.Website.Integration
     /// <summary>
     /// A class representing a factory for creating instances of the application.
     /// </summary>
-    public class TestServerFixture : WebApplicationFactory<Startup>, ITestOutputHelperAccessor
+    public class TestServerFixture : WebApplicationFactory<Services.IToolsService>, ITestOutputHelperAccessor
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="TestServerFixture"/> class.
